SRGAN parallel version
I am sure about SRResNet thanks to 
https://github.com/KJLiu/SRGAN-tensorflow/commit/0bb334721ea21df5c3d09780d2ef88b48c9cf0f7
I have extended this to SRGAN function but I am not very sure. Because in GAN, discriminator and generator train in turns and for distribution amongst tower generator should wait for discriminator. I have used optimizer in discriminator 2 times once inside the tower and once outside with average gradients.   Would be glad to get some open source help or comments  if not correct.